requirements.txt
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
jupyter==1.0.0
pymc==5.10.3
jax==0.4.23
hamiltorch @ git+https://github.com/AdamCobb/hamiltorch
laplace-torch==0.1a2
wandb==0.17.0
pyro-ppl==1.9.1
plotly==5.22.0
kaleido==0.2.1